Further expand "not from cwd" test re: cmd.exe
On Python versions for which python/cpython#101283 is not patched,
using Popen with shell=True can find cmd.exe in the current
directory on Windows, in the rare case that the ComSpec environment
variable is not defined.

This is not necessarily worth addressing, because it is a bug in
CPython rahter than GitPython, because that bug has been patched,
and because it is very rare that ComSpec is undefined. However:

- Changing the code to avoid it would also make that code simpler.
- Patched versions of Python <=3.9 don't have python.org builds.

This commit just expands the test to add cases where a repository
also has a file cmd.exe and where ComSpec is undefined, showing
that this case is not covered.