Omit CWD in search for bash.exe to run hooks on Windows
This uses the same NoDefaultCurrentDirectoryInExePath technique for
the Popen call in git.index.fun.run_commit_hook on Windows as is
already used in git.cmd.Git.execute.

The code is simpler in run_commit_hook because a shell is never
used to run bash.exe. (bash.exe is itself a shell, but we never
run it *via* a shell by passing shell=True to Popen.) Nonetheless,
it may make sense to extract out a helper function both can call.
This commit does not do that, so there is some code duplication.