Tweak formatting for `@pytest.mark.parametrize`
This removes the "fmt: off" / "fmt: on" directives around the
`@pytest.mark.parametrize` decoration on test_blob_filter, and
reformats it with black, for consistency with other such
decorations.

The style used there, *if* it could be specified as a rule and thus
used without "fmt:" directives, may be nicer than how black formats
multi-line mark decorations. However, since that decoration was
written, there have been a number of other such decorations, which
have been in black style.

This also removes the only (or only remaining?) "fmt:" directive in
the codebase. As such, it should possibly have been done in #1760.