Fix incompletely revised Iterable/IterableObj docstrings
Three wrong references in docstrings to list_items methods had
accidentally remained: one that was meant to be removed altogether,
and two that were meant to be references to iter_items.

This completes those changes, correcting the errors from
dfee31f (#1780) and 2b768c7 (#1785) so the docstrings make sense
and references to further information are to places with that
information.