In handle_process_output don't forward finalizer result
The git.cmd.handle_process_output function is non-public
(git.cmd.__all__ only lists Git) but used throughout GitPython and
referenced in the git.util.RemoteProgress.new_message_handler
docstring. Its finalizer argument is annotated to accept an
optional callable that always returns None. It is always used this
way in GitPython and RemoteProcess.new_message_handler is annotated
accordingly. However, the handle_process_output docstring and
implementation had documented the return value as the result of the
finalizer, and the implementation had forwarded that result if
passed.

This modifies the docstring and implementation to disregard any
result, in accordance with the everywhere-annotated assumption that
the finalizer is conceptually void and the absence of any code in
GitPython that uses the result of calling handle_process_output.
None is now implicitly returned, simplifying the implementation and
bringing it and the docstring in line with annotationd and usage.

This would be a breaking change if done on a public function, but
because handle_process_output is nonpublic (and documentation for
it is omitted by Sphinx, so users probably don't wrongly think it
is public), I believe this is safe and non-breaking.