Remove git.util.NullHandler
git.util has an __all__ attribute, which does not list NullHandler.
As NullHandler is not otherwise documented to be public, removing
it is not a breaking change, and there is no need for a deprecation
period or to wait until the next major version of GitPython.

See 741dfd3 for information about its history.