Don't add `NullHandler`s
This stops adding `NullHandler` instances to GitPython's loggers.
As noted in #1806, when they were added in #300 this prevented
errors when GitPython logged messages and logging was not enabled,
but since Python 3.2 there is a logger of last resort providing a
nicer default behavior of showing the messages. (They are still
shown with better formatting if logging is configured, even if
just done with logging.basicConfig(), so applications should still
typically configure logging.)