Use Python 3.9.16 on Cygwin CI
The latest currently packaged version of Python 3.9 for Cygwin is
3.9.18 (provided by the Cygwin package python39 at version
3.9.18-1). That version, at least as we are using it, has a problem
where pip has begun to block on some PyPI package downloads.

In 73ebcfa (EliahKagan#2), I worked around this problem by
downgrading the minor version of Python to 3.8. But it is better to
use 3.9 if we can, since it is currently the latest minor version
of Python in the Cygwin repositories, and also because (relating to
that) it is used more often, and thus probably used more often with
GitPython, than 3.8.

This upgrades Python on Cygwin but not all the way. It upgrades it
to the latest (or latest currently available) patch version of 3.9
packaged for Cygwin of those that strictly precede 3.9.18 where the
problem occurs. That version is 3.9.16, provided by the Cygwin
package python39 at version 3.9.16-1.

This version may eventually no longer be available for download
from Cygwin's repositories, so hopefully a real solution or better
workaround will be found by then, or perhaps a future update to the
package itself will fix the problem.

I also tried some more other stuff since finding 3.8 to work in
73ebcfa. Changes since then are listed below. They can be examined
in full detail, with individual diffs and CI results, at
https://github.com/EliahKagan/GitPython/pull/3.

* Try Python 3.9 with other details the same

Changing it to Python 3.8 worked, but I want to check that it was
actually the use of Python 3.8, rather than other seemingly small
changes made to support using Python 3.8, that made the difference.

* Revert "Try Python 3.9 with other details the same"

This reverts commit b55cbfbb31376e9da4ba6a64ac11cd791718ff8d.

* Try 3.9 again, with both python39=3.9.16-1 python39-pip=23.0.1-1

* Back to 3.8; try another GitHub Action

Python 3.8 worked with cygwin-install-action, but I want to make
the change to setup-cygwin by itself first before trying it with
3.9, in case I am using setup-cygwin incorrectly.

* Try 3.9 with this setup-cygwin action

* Try pinning with setup-cygwin

* Try not pinning, but no -U for PyPA, with setup-cygwin

Pinning and skipping -U for PyPA packages worked. Let's see if it
was really pinning that made the difference.

* Try pinning just python39-pip

Pinning works, and merely omitting the -U for PyPA package doesn't.
Examining the output of runs that used install-cygwin-action and
attemped pinning Cygwin package versions shows newer versions were
installed, whereas pinning is really happening with setup-cygwin.

This tries pinning just the Cygwin package for pip, rather than for
Python 3.9. I don't expect this to work.

* Try pinning just python39=3.9.16-1

And not pip, but this does not add back the -U for PyPA yet.

* Add back -U for PyPA packages

* Try pinning python39=3.9.16-1 with old action/everything

This is extremely unlikely to work, I just want to check.

* Try just setup-cygwin and pinning python39=3.9.16-1

That is, this puts back all the other stuff the way it was on the
main branch when the breakage occurred, besides changing from
cygwin-install-action to setup-cygwin to make pinning work and
using it to get version 3.9.16-1 of the Cygwin python39 package.